Refactor file part manipulation into individual classes
Rather than duplicate the same code to verify and get info about files
throughout the class, I've chopped it all into a class hierarchy for each
different type of file (kernel, initrd, defaults).  This makes it a bit
easier to reason about what's going on, and will also make life a bit easier
in the future when we start supporting more varieties of firmware file.